+++ apps/docs/app/(home)/layout.tsx
@@ -1,11 +1,11 @@
 import { Layout } from 'fumadocs-ui/layout';
 import type { ReactNode } from 'react';
-import { layoutOptions } from '@/app/layout.config';
+import { baseOptions } from '@/app/layout.config';
 
 export default function HomeLayout({
   children,
 }: {
   children: ReactNode;
 }): React.ReactElement {
-  return <Layout {...layoutOptions}>{children}</Layout>;
+  return <Layout {...baseOptions}>{children}</Layout>;
 }