+++ packages/ui/src/components/layout/nav.tsx
@@ -1,13 +1,11 @@
 'use client';
-import { MenuIcon, MoreVertical, SearchIcon, X } from 'lucide-react';
+import { MoreVertical } from 'lucide-react';
 import Link from 'fumadocs-core/link';
-import { SidebarTrigger } from 'fumadocs-core/sidebar';
-import { type ReactNode, useCallback, useEffect, useState } from 'react';
+import { type ReactNode, useEffect, useState } from 'react';
 import { usePathname } from 'next/navigation';
 import { cn } from '@/utils/cn';
 import { useSearchContext } from '@/contexts/search';
-import { useI18n } from '@/contexts/i18n';
-import { ThemeToggle } from '@/components/theme-toggle';
+import { ThemeToggle } from '@/components/layout/theme-toggle';
 import {
   Popover,
   PopoverContent,
@@ -15,7 +13,12 @@ import {
 } from '@/components/ui/popover';
 import { buttonVariants } from '@/theme/variants';
 import type { LinkItemType } from '@/layout';
-import { LinkItem } from './link-item';
+import {
+  LargeSearchToggle,
+  SearchToggle,
+} from '@/components/layout/search-toggle';
+import { useI18n } from '@/contexts/i18n';
+import { LinkItem } from '../link-item';
 
 export interface NavProps {
   title?: ReactNode;
@@ -25,10 +28,6 @@ export interface NavProps {
    * @defaultValue '/'
    */
   url?: string;
-
-  items: LinkItemType[];
-
-  enableSidebar: boolean;
 
   /**
    * Show/hide search toggle
@@ -50,10 +49,9 @@ export function Nav({
   url = '/',
   items,
   transparentMode = 'none',
-  enableSidebar,
   enableSearch = true,
   children,
-}: NavProps): React.ReactElement {
+}: NavProps & { items: LinkItemType[] }): React.ReactElement {
   const search = useSearchContext();
   const [transparent, setTransparent] = useState(transparentMode !== 'none');
 
@@ -98,27 +96,14 @@ export function Nav({
             />
           ))}
         <div className="flex flex-1 flex-row items-center justify-end md:gap-2">
-          {enableSearch && search.enabled ? <SearchToggle /> : null}
-          <ThemeToggle className="max-lg:hidden" />
-          {enableSidebar ? (
-            <SidebarTrigger
-              aria-label="Toggle Sidebar"
-              className={cn(
-                buttonVariants({
-                  size: 'icon',
-                  color: 'ghost',
-                  className: 'group md:hidden',
-                }),
-              )}
-            >
-              <MenuIcon className="group-data-[open=true]:hidden" />
-              <X className="hidden group-data-[open=true]:block" />
-            </SidebarTrigger>
+          {enableSearch && search.enabled ? (
+            <>
+              <SearchToggle className="md:hidden" />
+              <LargeSearchToggle className="w-full max-w-[240px] max-md:hidden" />
+            </>
           ) : null}
-          <LinksMenu
-            items={items}
-            className={cn('lg:hidden', enableSidebar && 'max-md:hidden')}
-          />
+          <ThemeToggle className="max-lg:hidden" />
+          <LinksMenu items={items} className="lg:hidden" />
           {items
             .filter((item) => item.type === 'secondary')
             .map((item, i) => (
@@ -136,6 +121,7 @@ interface LinksMenuProps extends React.ButtonHTMLAttributes<HTMLButtonElement> {
 
 function LinksMenu({ items, ...props }: LinksMenuProps): React.ReactElement {
   const [open, setOpen] = useState(false);
+  const { text } = useI18n();
   const pathname = usePathname();
 
   useEffect(() => {
@@ -161,52 +147,12 @@ function LinksMenu({ items, ...props }: LinksMenuProps): React.ReactElement {
           <LinkItem key={i} item={item} on="menu" />
         ))}
         <div className="flex flex-row items-center justify-between px-2 py-1">
-          <p className="font-medium">Theme</p>
+          <p className="font-medium text-muted-foreground">
+            {text.chooseTheme}
+          </p>
           <ThemeToggle />
         </div>
       </PopoverContent>
     </Popover>
-  );
-}
-
-function SearchToggle(): React.ReactElement {
-  const { hotKey, setOpenSearch } = useSearchContext();
-  const { text } = useI18n();
-  const onClick = useCallback(() => {
-    setOpenSearch(true);
-  }, [setOpenSearch]);
-
-  return (
-    <>
-      <button
-        type="button"
-        className={cn(
-          buttonVariants({
-            size: 'icon',
-            color: 'ghost',
-            className: 'md:hidden',
-          }),
-        )}
-        aria-label="Open Search"
-        onClick={onClick}
-      >
-        <SearchIcon />
-      </button>
-      <button
-        type="button"
-        className="inline-flex w-full max-w-[240px] items-center gap-2 rounded-full border bg-secondary/50 p-1.5 text-sm text-muted-foreground transition-colors hover:bg-accent hover:text-accent-foreground max-md:hidden"
-        onClick={onClick}
-      >
-        <SearchIcon className="ms-1 size-4" />
-        {text.search}
-        <div className="ms-auto inline-flex gap-0.5 text-xs">
-          {hotKey.map((k, i) => (
-            <kbd key={i} className="rounded-md border bg-background px-1.5">
-              {k.display}
-            </kbd>
-          ))}
-        </div>
-      </button>
-    </>
   );
 }