+++ packages/ui/src/contexts/sidebar.tsx
@@ -1,18 +1,30 @@
-import { createContext, useContext, useState, useMemo } from 'react';
+import {
+  createContext,
+  useContext,
+  useState,
+  useMemo,
+  useRef,
+  type MutableRefObject,
+  useEffect,
+} from 'react';
+import { usePathname } from 'next/navigation';
 import { SidebarProvider as BaseProvider } from 'fumadocs-core/sidebar';
 
-interface SidebarCollapseContext {
+interface SidebarContext {
   open: boolean;
   setOpen: (v: boolean) => void;
   collapsed: boolean;
   setCollapsed: (v: boolean) => void;
+
+  /**
+   * When set to true, close the sidebar on redirection
+   */
+  closeOnRedirect: MutableRefObject<boolean>;
 }
 
-const SidebarContext = createContext<SidebarCollapseContext | undefined>(
-  undefined,
-);
+const SidebarContext = createContext<SidebarContext | undefined>(undefined);
 
-export function useSidebar(): SidebarCollapseContext {
+export function useSidebar(): SidebarContext {
   const ctx = useContext(SidebarContext);
   if (!ctx) throw new Error('Missing root provider');
   return ctx;
@@ -23,8 +35,18 @@ export function SidebarProvider({
 }: {
   children: React.ReactNode;
 }): React.ReactElement {
+  const closeOnRedirect = useRef(false);
   const [open, setOpen] = useState(false);
   const [collapsed, setCollapsed] = useState(false);
+
+  const pathname = usePathname();
+
+  useEffect(() => {
+    if (closeOnRedirect.current) {
+      setOpen(false);
+      closeOnRedirect.current = false;
+    }
+  }, [pathname]);
 
   return (
     <SidebarContext.Provider
@@ -34,6 +56,7 @@ export function SidebarProvider({
           setOpen,
           collapsed,
           setCollapsed,
+          closeOnRedirect,
         }),
         [open, collapsed],
       )}