+++ packages/ui/src/layout.tsx
@@ -1,16 +1,24 @@
 import type { PageTree } from 'fumadocs-core/server';
 import type { ReactNode, HTMLAttributes } from 'react';
-import type { NavProps } from './components/nav';
+import Link from 'next/link';
+import { cn } from '@/utils/cn';
+import type { NavProps } from './components/layout/nav';
 import { replaceOrDefault } from './utils/shared';
-import { cn } from './utils/cn';
-import type { SidebarProps } from './components/sidebar';
+import type { SidebarProps } from './components/layout/sidebar';
+
+declare const { Nav }: typeof import('./layout.client');
 
 declare const {
-  Nav,
   TreeContextProvider,
   DynamicSidebar,
+  SubNav,
   Sidebar,
-}: typeof import('./layout.client');
+  ThemeToggle,
+}: typeof import('./docs.client');
+
+declare const {
+  LanguageToggle,
+}: typeof import('./components/layout/language-toggle');
 
 type ActiveType = 'none' | 'url' | 'nested-url';
 
@@ -48,15 +56,9 @@ export type LinkItemType =
       external?: boolean;
     };
 
-interface NavOptions
-  extends Omit<NavProps, 'enableSidebar' | 'collapsibleSidebar' | 'items'> {
+interface NavOptions extends Omit<NavProps, 'items'> {
   enabled: boolean;
   component: ReactNode;
-
-  /**
-   * GitHub url displayed on the navbar
-   */
-  githubUrl: string;
 }
 
 interface SidebarOptions extends Omit<SidebarProps, 'items'> {
@@ -66,13 +68,18 @@ interface SidebarOptions extends Omit<SidebarProps, 'items'> {
 }
 
 export interface BaseLayoutProps {
+  /**
+   * GitHub url
+   */
+  githubUrl?: string;
+
   links?: LinkItemType[];
   /**
    * Replace or disable navbar
    */
   nav?: Partial<NavOptions>;
 
-  children: ReactNode;
+  children?: ReactNode;
 }
 
 export interface DocsLayoutProps extends BaseLayoutProps {
@@ -81,20 +88,28 @@ export interface DocsLayoutProps extends BaseLayoutProps {
   sidebar?: Partial<SidebarOptions>;
 
   containerProps?: HTMLAttributes<HTMLDivElement>;
+
+  /**
+   * Enable Language Switch
+   *
+   * @defaultValue false
+   */
+  i18n?: boolean;
 }
 
 export function Layout({
   nav = {},
   links = [],
+  githubUrl,
   children,
 }: BaseLayoutProps): React.ReactElement {
-  const finalLinks = getLinks(links, nav.githubUrl);
+  const finalLinks = getLinks(links, githubUrl);
 
   return (
     <>
       {replaceOrDefault(
         nav,
-        <Nav items={finalLinks} enableSidebar={false} {...nav}>
+        <Nav items={finalLinks} {...nav}>
           {nav.children}
         </Nav>,
       )}
@@ -104,49 +119,56 @@ export function Layout({
 }
 
 export function DocsLayout({
-  nav = {},
+  nav,
+  githubUrl,
   sidebar = {},
   links = [],
   containerProps,
   tree,
+  i18n = false,
   children,
 }: DocsLayoutProps): React.ReactElement {
   const sidebarEnabled = sidebar.enabled ?? true;
   const sidebarCollapsible = sidebarEnabled && (sidebar.collapsible ?? true);
-  const finalLinks = getLinks(links, nav.githubUrl);
+  const finalLinks = getLinks(links, githubUrl);
+
+  const Aside = sidebarCollapsible ? DynamicSidebar : Sidebar;
 
   return (
     <TreeContextProvider tree={tree}>
-      {replaceOrDefault(
-        nav,
-        <Nav items={finalLinks} enableSidebar={sidebarEnabled} {...nav}>
-          {nav.children}
-        </Nav>,
-      )}
+      {replaceOrDefault(nav, <SubNav {...nav} />)}
       <div
         {...containerProps}
-        className={cn(
-          'mx-auto flex w-full max-w-container flex-row gap-2 xl:gap-6',
-          containerProps?.className,
-        )}
+        className={cn('flex flex-1 flex-row', containerProps?.className)}
       >
         {replaceOrDefault(
           sidebar,
-          sidebarCollapsible ? (
-            <DynamicSidebar
-              items={finalLinks}
-              defaultOpenLevel={sidebar.defaultOpenLevel}
-              banner={sidebar.banner}
-              footer={sidebar.footer}
-            />
-          ) : (
-            <Sidebar
-              items={finalLinks}
-              defaultOpenLevel={sidebar.defaultOpenLevel}
-              banner={sidebar.banner}
-              footer={sidebar.footer}
-            />
-          ),
+          <Aside
+            items={finalLinks}
+            defaultOpenLevel={sidebar.defaultOpenLevel}
+            banner={
+              <>
+                <Link
+                  href={nav?.url ?? '/'}
+                  className="inline-flex items-center gap-2.5 border-b pb-4 font-semibold max-md:hidden"
+                >
+                  {nav?.title}
+                </Link>
+                {sidebar.banner}
+              </>
+            }
+            bannerProps={{
+              className: cn(!sidebar.banner && 'max-md:hidden'),
+            }}
+            footer={
+              <>
+                <ThemeToggle className="me-auto" />
+                {sidebar.footer}
+                {i18n ? <LanguageToggle /> : null}
+              </>
+            }
+            components={sidebar.components}
+          />,
         )}
 
         {children}